- Created Put item endpoint - Created components folder to split js and html code - Started working on having an empty item name error returned
