+++ simple-shopping-cart-source/Controllers/ItemController.cs
@@ -1,8 +1,5 @@
-﻿using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
+﻿using System.Linq;
 using System.Web.Http;
-using System.Web.UI.WebControls;
 using simple_shopping_cart_source.Models;
 
 namespace simple_shopping_cart_source.Controllers
@@ -59,7 +56,6 @@ namespace simple_shopping_cart_source.Controllers
             if (changedItem == null) return NotFound();
 
             changedItem.Checked = item.Checked;
-            //return Ok(changedItem);
 
             return Ok(shoppingList);
         }
