+++ simple-shopping-cart-source/simple-shopping-cart-source.csproj
@@ -89,9 +89,13 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Content Include="Components\AddItem\addItem.css" />
+    <Content Include="Components\AddItem\addItem.html" />
+    <Content Include="Components\AddItem\addItem.js" />
     <Content Include="Global.asax" />
     <None Include="Scripts\jquery-3.4.1.intellisense.js" />
     <Content Include="Images\favicon-16x16.png" />
+    <Content Include="index.html" />
     <Content Include="Scripts\app.js" />
     <Content Include="Scripts\jquery-3.4.1.js" />
     <Content Include="Scripts\jquery-3.4.1.min.js" />
